requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
dj-database-url==0.5.0
django-crispy-forms==1.13.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
multidict==5.1.0
psycopg2==2.9.1
pycparser==2.20
pyfiglet==0.8.post1
pytz==2021.1
requests==2.26.0
rstr==3.1.0
six==1.16.0
sqlparse==0.4.2
telepot==12.7
typing-extensions==3.10.0.2
urllib3==1.26.7
user-agent==0.1.9
uuid==1.30
whitenoise==5.3.0
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0

